requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
fire==0.7.0
fonttools==4.56.0
lxml==5.3.1
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pdf2docx==0.5.8
pillow==11.1.0
PyMuPDF==1.25.4
pytesseract==0.3.13
python-docx==1.1.2
python-magic==0.4.27
python-multipart==0.0.20
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.13.0
tzdata==2025.2
